+++ src/main/java/io/github/pustike/inject/impl/DefaultInjectionPointLoader.java
@@ -15,17 +15,9 @@
  */
 package io.github.pustike.inject.impl;
 
-import java.lang.annotation.Annotation;
-import java.lang.reflect.Array;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Executable;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
-import java.lang.reflect.ParameterizedType;
-import java.lang.reflect.Type;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
@@ -33,10 +25,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import javax.inject.Inject;
-import javax.inject.Provider;
-import javax.inject.Qualifier;
 
-import io.github.pustike.inject.BindingKey;
 import io.github.pustike.inject.bind.InjectionPoint;
 import io.github.pustike.inject.bind.InjectionPointLoader;
 
@@ -63,18 +52,15 @@ public final class DefaultInjectionPointLoader implements InjectionPointLoader {
     public static List<InjectionPoint<Object>> doCreateInjectionPoints(final Class<?> targetClass) {
         List<InjectionPoint<Object>> injectionPointList = new LinkedList<>();
         Set<Integer> visitedMethodHashCodeSet = new HashSet<>();
-        for (Class clazz = targetClass; clazz != null; clazz = clazz.getSuperclass()) {
-            if (clazz == Object.class) {
-                continue;
-            }
+        for (Class clazz = targetClass; clazz != Object.class; clazz = clazz.getSuperclass()) {
             int index = 0, staticIndex = 0;
             Field[] fields = clazz.getDeclaredFields();
             for (Field field : fields) {
                 if (field.getDeclaredAnnotation(Inject.class) != null) {
                     if (Modifier.isStatic(field.getModifiers())) {
-                        injectionPointList.add(staticIndex++, createInjectionPoint(field));
+                        injectionPointList.add(staticIndex++, new FieldInjectionPoint<>(field));
                     } else {
-                        injectionPointList.add(index, createInjectionPoint(field));
+                        injectionPointList.add(index, new FieldInjectionPoint<>(field));
                     }
                     index++;
                 }
@@ -88,77 +74,15 @@ public final class DefaultInjectionPointLoader implements InjectionPointLoader {
                 visitedMethodHashCodeSet.add(hashCode);
                 if (method.getDeclaredAnnotation(Inject.class) != null) {
                     if (Modifier.isStatic(method.getModifiers())) {
-                        injectionPointList.add(staticIndex++, createInjectionPoint(method));
+                        injectionPointList.add(staticIndex++, new ExecutableInjectionPoint<>(method));
                     } else {
-                        injectionPointList.add(index, createInjectionPoint(method));
+                        injectionPointList.add(index, new ExecutableInjectionPoint<>(method));
                     }
                     index++;
                 }
             }
         }
         return injectionPointList;
-    }
-
-    static <T> InjectionPoint<T> createInjectionPoint(Class<? extends T> targetType) {
-        @SuppressWarnings("unchecked")
-        Constructor<T>[] constructors = (Constructor<T>[]) targetType.getDeclaredConstructors();
-        if (constructors.length == 0) {
-            throw new RuntimeException("No constructors available for type: " + targetType);
-        }
-        Constructor<T> defaultConstructor = null;
-        for (Constructor<T> constructor : constructors) {
-            if (constructor.isAnnotationPresent(Inject.class)) {
-                return createInjectionPoint(constructor);
-            }
-            if (constructor.getParameterCount() == 0) {
-                defaultConstructor = constructor;
-            }
-        }
-        if (defaultConstructor == null) {
-            throw new RuntimeException("No constructors available for type: " + targetType);
-        }
-        return createInjectionPoint(defaultConstructor);
-    }
-
-    static <T> InjectionPoint<T> createInjectionPoint(Executable executable) {
-        Type[] parameterTypes = executable.getGenericParameterTypes();
-        Annotation[][] annotations = executable.getParameterAnnotations();
-        BindingKey<T>[] parameterBindingKeys = createParameterBindingKeys(parameterTypes, annotations);
-        Boolean[] nullableParams = new Boolean[parameterTypes.length];
-        Arrays.setAll(nullableParams, i -> allowsNullValue(annotations[i]));
-        return new ExecutableInjectionPoint<>(executable, parameterBindingKeys, nullableParams);
-    }
-
-    @SuppressWarnings("unchecked")
-    private static <T> InjectionPoint<T> createInjectionPoint(Field field) {
-        Annotation[] annotations = field.getAnnotations();
-        BindingKey<T> targetKey = createBindingKey(field.getGenericType(), annotations);
-        return new FieldInjectionPoint<>(field, targetKey, allowsNullValue(annotations));
-    }
-
-    @SuppressWarnings("unchecked")
-    private static <T> BindingKey<T>[] createParameterBindingKeys(Type[] parameterTypes, Annotation[][] annotations) {
-        BindingKey<T>[] bindingKeys = (BindingKey<T>[]) Array.newInstance(BindingKey.class, parameterTypes.length);
-        Arrays.setAll(bindingKeys, i -> createBindingKey(parameterTypes[i], annotations[i]));
-        return bindingKeys;
-    }
-
-    @SuppressWarnings("unchecked")
-    static <T> BindingKey<T> createBindingKey(Type genericType, Annotation[] annotations) {
-        Type rawType = genericType instanceof ParameterizedType ?
-                ((ParameterizedType) genericType).getRawType() : genericType;
-        boolean isMultiBinder = List.class.equals(rawType) || Collection.class.equals(rawType)
-                || Iterable.class.equals(rawType);
-        if (isMultiBinder) {
-            genericType = getTypeArgument(genericType);
-            rawType = genericType instanceof ParameterizedType ?
-                    ((ParameterizedType) genericType).getRawType() : genericType;
-        }
-        boolean isProviderType = Provider.class.equals(rawType);
-        Class<T> bindingType = (Class<T>) (isProviderType ? getTypeArgument(genericType) : rawType);
-        BindingKey<T> bindingKey = BindingKey.of(bindingType, getQualifierAnnotation(annotations));
-        bindingKey = isMultiBinder ? (BindingKey<T>) bindingKey.toListType() : bindingKey;
-        return isProviderType ? (BindingKey<T>) bindingKey.toProviderType() : bindingKey;
     }
 
     private static int computeHashCode(Class clazz, Method method) {
@@ -174,34 +98,5 @@ public final class DefaultInjectionPointLoader implements InjectionPointLoader {
             hashCode = 31 * hashCode + clazz.hashCode(); // private method
         }
         return hashCode;
-    }
-
-    private static Annotation getQualifierAnnotation(Annotation[] annotations) {
-        for (Annotation annotation : annotations) {
-            if (annotation.annotationType().isAnnotationPresent(Qualifier.class)) {
-                return annotation;
-            }
-        }
-        return null;
-    }
-
-    private static Type getTypeArgument(Type genericType) {
-        if (genericType instanceof ParameterizedType) {
-            Type[] typeArgs = ((ParameterizedType) genericType).getActualTypeArguments();
-            if (typeArgs != null && typeArgs.length == 1) {
-                return typeArgs[0];
-            }
-        }
-        return genericType;
-    }
-
-    private static boolean allowsNullValue(Annotation[] annotations) {
-        for (Annotation a : annotations) {
-            Class<? extends Annotation> type = a.annotationType();
-            if ("Nullable".equals(type.getSimpleName())) {
-                return true;
-            }
-        }
-        return false;
     }
 }